requirements.txt
asgiref>=3.2.3
astroid>=2.3.3
cffi>=1.13.2
colorama>=0.4.3
cryptography>=2.8
dj-database-url>=0.5.0
Django>=3.0.2
django-cors-headers>=3.2.1
django-heroku>=0.3.1
django-rest-knox>=4.1.0
djangorestframework>=3.11.0
gunicorn>=20.0.4
isort>=4.3.21
lazy-object-proxy>=1.4.3
mccabe>=0.6.1
Pillow>=7.0.0
psycopg2>=2.8.4
pycparser>=2.19
pylint>=2.4.4
pytz>=2019.3
six>=1.14.0
sqlparse>=0.3.0
whitenoise>=5.0.1
wrapt>=1.11.2
